requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic=1.15.1=pypi_0
aniso8601=10.0.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
attrs=25.3.0=pypi_0
blinker=1.9.0=pypi_0
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pypi_0
filelock=3.13.1=pypi_0
flasgger=0.9.7.1=pypi_0
flask=3.1.0=pypi_0
flask-cors=5.0.1=pypi_0
flask-login=0.6.3=pypi_0
flask-migrate=4.1.0=pypi_0
flask-restful=0.3.10=pypi_0
flask-restx=1.3.0=pypi_0
flask-sqlalchemy=3.1.1=pypi_0
fonttools=4.56.0=pypi_0
fpdf=1.7.2=pypi_0
fsspec=2024.6.1=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
iniconfig=2.0.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
llvmlite=0.43.0=pypi_0
lxml=5.3.1=pypi_0
mako=1.3.9=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
mistune=3.1.3=pypi_0
more-itertools=10.6.0=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=2.0.2=pypi_0
openai=1.65.3=pypi_0
openai-whisper=20240930=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.0.16=h3f729d1_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py39haa95532_0
pluggy=1.5.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pyparsing=3.2.3=pypi_0
pypdf2=3.0.1=pypi_0
pytesseract=0.3.13=pypi_0
pytest=8.3.5=pypi_0
python=3.9.21=h8205438_1
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.24.0=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=76.0.0=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.38=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2024.8.30=pypi_0
tiktoken=0.9.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.6.0=pypi_0
torchaudio=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.3.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39haa95532_0
zipp=3.21.0=pypi_0
